% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{add_proportions}
\alias{add_proportions}
\title{Add ratio of severe cases + deaths to clinical cases to df}
\usage{
add_proportions(dt)
}
\arguments{
\item{dt}{case output at country level}
}
\description{
Add ratio of severe cases + deaths to clinical cases to df
}
