% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_doses}
\alias{aggregate_doses}
\title{Aggregate dose outputs}
\usage{
aggregate_doses(doses_full)
}
\arguments{
\item{doses_full}{dose input}
}
\value{
aggregated dose output
}
\description{
Aggregate dose outputs
}
