% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_outputs}
\alias{aggregate_outputs}
\title{Aggregate VIMC outputs up to the country level}
\usage{
aggregate_outputs(dt, pop)
}
\arguments{
\item{dt}{data frame with site-level outputs to aggregate}

\item{pop}{VIMC population input for country level (age- and year-specific)}
}
\value{
data frame with aggregated outputs up to the country level
}
\description{
Aggregate VIMC outputs up to the country level
}
