% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_site}
\alias{analyse_site}
\title{model vaccine impact}
\usage{
analyse_site(site, site_data, vimc_input, baseline_outputs)
}
\arguments{
\item{site}{analysis map with input parameters}

\item{site_data}{site data}

\item{vimc_input}{vimc_input}

\item{baseline_outputs}{baseline_outputs}
}
\value{
modelled + processed output
}
\description{
model vaccine impact
}
