% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{append_low_transmission_sites}
\alias{append_low_transmission_sites}
\title{Append output from low-transmission admin units to full dataset}
\usage{
append_low_transmission_sites(low_transmission, intvn)
}
\arguments{
\item{low_transmission}{low-transmission output}

\item{intvn}{output from models}
}
\description{
Append output from low-transmission admin units to full dataset
}
