% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{calculate_cases_averted}
\alias{calculate_cases_averted}
\title{Calculate cases averted}
\usage{
calculate_cases_averted(intvn_output, baseline_output)
}
\arguments{
\item{intvn_output}{}
}
\description{
Calculate cases averted
}
