% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_file.R
\name{check_eir}
\alias{check_eir}
\title{Check that the EIR of the site of interest is not zero}
\usage{
check_eir(site)
}
\arguments{
\item{site}{site file}

\item{iso3c}{Country code}
}
\value{
sites component of site file with only sites that will be modelled
}
\description{
Check that the EIR of the site of interest is not zero
}
