% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{check_not_a_rerun}
\alias{check_not_a_rerun}
\title{Return a list of the reports that are not a rerun based on orderly metadata}
\usage{
check_not_a_rerun(report_name, map, date_time)
}
\arguments{
\item{report_name}{name of orderly report to extract metadata from}

\item{map}{parameter map for reports you would like to crossreference}

\item{date_time}{check if reports completed after a certain date, format YYYYMMDDHHMMSS}
}
\description{
Return a list of the reports that are not a rerun based on orderly metadata
}
