% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{compile_and_save}
\alias{compile_and_save}
\title{save final outputs from workflow}
\usage{
compile_and_save(description, scen)
}
\arguments{
\item{description}{description of runs to pull}

\item{scen}{scenario to save}
}
\description{
save final outputs from workflow
}
