% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{compile_dose_outputs}
\alias{compile_dose_outputs}
\title{Pull final outputs from workflow}
\usage{
compile_dose_outputs(descrip, iso)
}
\arguments{
\item{descrip}{description of runs to pull}
}
\description{
Pull final outputs from workflow
}
