% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{compile_final_outputs}
\alias{compile_final_outputs}
\title{Pull final outputs from workflow}
\usage{
compile_final_outputs(descrip)
}
\arguments{
\item{descrip}{description of runs to pull}
}
\description{
Pull final outputs from workflow
}
