% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{expand_intervention_coverage}
\alias{expand_intervention_coverage}
\title{expand intervention years out to terminal year of forecast using scene package}
\usage{
expand_intervention_coverage(site, terminal_year)
}
\arguments{
\item{site}{site file}

\item{terminal_year}{terminal year of forecast}
}
\value{
site file with extrapolated coverage values out to terminal year
}
\description{
expand intervention years out to terminal year of forecast using scene package
}
