% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{expand_life_expectancy}
\alias{expand_life_expectancy}
\title{expand VIMC life expectancy data frame to single yea}
\usage{
expand_life_expectancy(le, iso3c)
}
\arguments{
\item{le}{VIMC life expectacy input}

\item{iso3c}{country code}
}
\description{
expand VIMC life expectancy data frame to single yea
}
