% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_file.R
\name{extract_site}
\alias{extract_site}
\title{Extract a single site input from a country site file}
\usage{
extract_site(site_file, site_name, ur)
}
\arguments{
\item{site_file}{Country site file}

\item{site_name}{name of site to extract}

\item{urbanicity}{urbanicity of site to extract}
}
\value{
Single site
}
\description{
Extract a single site input from a country site file
}
