% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{format_descriptive_data}
\alias{format_descriptive_data}
\title{format descriptive data for diagnostic report}
\usage{
format_descriptive_data()
}
\description{
format descriptive data for diagnostic report
}
