% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{format_outputs}
\alias{format_outputs}
\title{format outputs for submission}
\usage{
format_outputs(dt, iso3c, site_name, ur, scenario, description)
}
\arguments{
\item{dt}{postprocessed output}

\item{iso3c}{country code}

\item{site_name}{name of site}

\item{ur}{urbanicity}

\item{scenario}{vaccine scenario}

\item{description}{reason for model run}
}
\description{
format outputs for submission
}
