% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{generate_parameter_map_for_next_report}
\alias{generate_parameter_map_for_next_report}
\title{before you run a report, check that the preceding reports have already been run}
\usage{
generate_parameter_map_for_next_report(report_name, parameter_map, day = NULL)
}
\arguments{
\item{report_name}{name of orderly report to check}

\item{parameter_map}{map of parameters for model runs to check against}

\item{day}{if you would only like to check reports completed after a certain day, MMDDYYYY format}
}
\value{
parameter map of reports that have not been run before
}
\description{
before you run a report, check that the preceding reports have already been run
}
