% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{get_dose_output}
\alias{get_dose_output}
\title{Pull site level processed output based on metadata input}
\usage{
get_dose_output(index, map, output_filepath)
}
\arguments{
\item{index}{observation in metadata df}

\item{map}{metadata df}

\item{output_filepath}{filepath where outputs live}
}
\description{
Pull site level processed output based on metadata input
}
