% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{get_raw_output}
\alias{get_raw_output}
\title{Pull site level dose output based on metadata input}
\usage{
get_raw_output(index, map, output_filepath)
}
\arguments{
\item{index}{observation in metadata df}

\item{map}{metadata df}

\item{output_filepath}{filepath where outputs live}
}
\description{
Pull site level dose output based on metadata input
}
