% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{make_analysis_map}
\alias{make_analysis_map}
\title{make an analysis map of input parameters for vaccine modelling run}
\usage{
make_analysis_map(site_df, test, run_all)
}
\arguments{
\item{site_df}{analysis map with input parameters}

\item{test}{boolean-- if true, only run analysis for two test sites. Good for quick tests of code functionality}

\item{run_all}{run all sites regardless of pfpr (boolean)}
}
\value{
analysis map to be used as an input for analyse_site
}
\description{
make an analysis map of input parameters for vaccine modelling run
}
