% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{parameterize_stochastic_run}
\alias{parameterize_stochastic_run}
\title{Change parameter set based off of stochastic draws}
\usage{
parameterize_stochastic_run(params, parameter_draw)
}
\arguments{
\item{params}{model input parameters}

\item{parameter_draw}{parameter draw number (0-50)}
}
\value{
parameters for stochastic run
}
\description{
Change parameter set based off of stochastic draws
}
