% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{pull_age_groups_time_horizon}
\alias{pull_age_groups_time_horizon}
\title{Make list of parameters that will apply to all model runs}
\usage{
pull_age_groups_time_horizon(quick_run)
}
\arguments{
\item{quick_run}{quick run setting}
}
\description{
Make list of parameters that will apply to all model runs
}
