% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{pull_doses_output}
\alias{pull_doses_output}
\title{pull dosage output}
\usage{
pull_doses_output(raw_output, processed_output)
}
\arguments{
\item{raw_output}{model output}

\item{processed_output}{name of site}
}
\description{
pull dosage output
}
