% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{pull_low_transmission_sites}
\alias{pull_low_transmission_sites}
\title{For vaccine scenarios, pull no-vaccination outputs for admin units < 10 \% PFPR}
\usage{
pull_low_transmission_sites(iso, processed_sites)
}
\arguments{
\item{processed_sites}{site level output}

\item{iso3c}{country code}
}
\description{
For vaccine scenarios, pull no-vaccination outputs for admin units < 10 \% PFPR
}
