% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{pull_most_recent_output}
\alias{pull_most_recent_output}
\title{Pull metadata from process country report}
\usage{
pull_most_recent_output(iso3c, description, quick_run)
}
\arguments{
\item{iso3c}{country code}

\item{description}{reason for model run}

\item{quick_run}{quick run setting}
}
\description{
Pull metadata from process country report
}
