% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{pull_other_case_metrics}
\alias{pull_other_case_metrics}
\title{Pull  figures for diagnostics}
\usage{
pull_other_case_metrics(site_data, intvn_results, bl_results)
}
\arguments{
\item{site_data}{}

\item{intvn_results}{}
}
\description{
Pull  figures for diagnostics
}
