% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{pull_outcomes_averted_per_100k_vacc}
\alias{pull_outcomes_averted_per_100k_vacc}
\title{Pull additional figures for diagnostics}
\usage{
pull_outcomes_averted_per_100k_vacc(intvn_output, baseline_output, doses)
}
\arguments{
\item{intvn_output}{}

\item{doses}{}
}
\description{
Pull additional figures for diagnostics
}
