% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pull_plotting_data}
\alias{pull_plotting_data}
\title{format plotting data for diagnostic report}
\usage{
pull_plotting_data(scenario)
}
\arguments{
\item{scenario}{vaccine scenario}
}
\description{
format plotting data for diagnostic report
}
