% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{recalibrate}
\alias{recalibrate}
\title{recalibration for certain countries (Guinea Bissau in this case, because elimination is occurring there)}
\usage{
recalibrate(input, site_data, yr)
}
\arguments{
\item{input}{modelling input}

\item{site_data}{site data which contains the MAP prevalence data}

\item{yr}{year to calibrate to MAP prevalence data}
}
\value{
site file with extrapolated coverage values out to terminal year
}
\description{
recalibration for certain countries (Guinea Bissau in this case, because elimination is occurring there)
}
