% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{reformat_output}
\alias{reformat_output}
\title{format output from analyse_site}
\usage{
reformat_output(output)
}
\arguments{
\item{output}{model output}
}
\description{
format output from analyse_site
}
