% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{remove_duplicate_reports}
\alias{remove_duplicate_reports}
\title{check if you have run this report before; if so remove from list of parameters to run
note you may want to rerun a report with the same parameters if you have changed source code;
in that case do not use this function}
\usage{
remove_duplicate_reports(report_name, parameter_map, day = NULL)
}
\arguments{
\item{report_name}{name of orderly report to check}

\item{parameter_map}{map of parameters for model runs to check against}

\item{day}{if you would only like to check reports completed after a certain day, MMDDYYYY format}
}
\value{
parameter map of reports that have not been run before
}
\description{
check if you have run this report before; if so remove from list of parameters to run
note you may want to rerun a report with the same parameters if you have changed source code;
in that case do not use this function
}
