% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_file.R
\name{remove_zero_eirs}
\alias{remove_zero_eirs}
\title{Check that the sites you seek to model have non-zero EIRs}
\usage{
remove_zero_eirs(iso3c, sites)
}
\arguments{
\item{iso3c}{Country code}

\item{sites}{sites component of site file}
}
\value{
sites component of site file with only sites that will be modelled
}
\description{
Check that the sites you seek to model have non-zero EIRs
}
