% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_baseline_model}
\alias{run_baseline_model}
\title{Run baseline model for the VIMC scenarios (no vaccination, 15 years of burnin + )}
\usage{
run_baseline_model(model_input)
}
\arguments{
\item{model_input}{list with input parameters and identifying info}
}
\value{
model output
}
\description{
Run baseline model for the VIMC scenarios (no vaccination, 15 years of burnin + )
}
