% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_local_reports}
\alias{run_local_reports}
\title{Loop through reports locally based on parameters in map}
\usage{
run_local_reports(map, report_name)
}
\arguments{
\item{map}{parameter map for reports}

\item{report_name}{name of orderly report}
}
\description{
Loop through reports locally based on parameters in map
}
