% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_model}
\alias{run_model}
\title{Run model for site of interest}
\usage{
run_model(model_input)
}
\arguments{
\item{model_input}{list with input parameters and identifying info}
}
\value{
model output
}
\description{
Run model for site of interest
}
