% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_scenario_model}
\alias{run_scenario_model}
\title{Run baseline model for the VIMC scenarios (no vaccination, 15 years of burnin + )}
\usage{
run_scenario_model(model_input, baseline_outputs)
}
\arguments{
\item{model_input}{list with input parameters and identifying info}
}
\value{
model output
}
\description{
Run baseline model for the VIMC scenarios (no vaccination, 15 years of burnin + )
}
