% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{scale_cases_deaths}
\alias{scale_cases_deaths}
\title{scale outputs based on cases from WMR from 2000-2020}
\usage{
scale_cases_deaths(dt, site_data, iso3c, scaling_data)
}
\arguments{
\item{dt}{case output at country level}

\item{site_data}{site file}

\item{scaling_data}{data used to scale (from no-vaccine scenario)}
}
\description{
scale outputs based on cases from WMR from 2000-2020
}
