% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{scale_par}
\alias{scale_par}
\title{Scale outputs based on difference in VIMC population at risk + site file population at risk}
\usage{
scale_par(processed_output, iso3c)
}
\arguments{
\item{processed_output}{processed output}

\item{iso3c}{country code}
}
\description{
Scale outputs based on difference in VIMC population at risk + site file population at risk
}
