% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{scale_population}
\alias{scale_population}
\title{scale admin-1 populations to be sum to country population input from VIMC}
\usage{
scale_population(site_data, site_name, ur, iso3c, vimc_pop)
}
\arguments{
\item{site_data}{site file}

\item{site_name}{name of site}

\item{ur}{urbanicity}

\item{iso3c}{country of interest}

\item{vimc_pop}{VIMC population input}
}
\description{
scale admin-1 populations to be sum to country population input from VIMC
}
