% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{site_postprocessing}
\alias{site_postprocessing}
\title{postprocess model output at the site level}
\usage{
site_postprocessing(id, dt)
}
\arguments{
\item{id}{identifier for unique site unit-- typically some combination of site name and urbanicity}

\item{dt}{raw model output}
}
\value{
postprocessed output at the site level
}
\description{
postprocess model output at the site level
}
