% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{submit_by_core}
\alias{submit_by_core}
\title{Submit jobs by core}
\usage{
submit_by_core(core, dt)
}
\arguments{
\item{core}{number of cores to request for job}

\item{dt}{parameter map for reports}
}
\description{
Submit jobs by core
}
