% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{subset_high_transmission_sites}
\alias{subset_high_transmission_sites}
\title{Function for a data request from Gates Foundation
Typically, in VIMC modelling vaccines are administered to all admin units above PFPR of 10\% in 2019
Raise threshold for extra analysis to 35\%}
\usage{
subset_high_transmission_sites(intvn, threshold, site_data)
}
\arguments{
\item{intvn}{site level output for intervention sites}

\item{threshold}{PFPR cutoff for vaccine administration}

\item{site_data}{site data}
}
\description{
Function for a data request from Gates Foundation
Typically, in VIMC modelling vaccines are administered to all admin units above PFPR of 10\% in 2019
Raise threshold for extra analysis to 35\%
}
