% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{update_coverage_values}
\alias{update_coverage_values}
\title{update vaccine coverage based on VIMC inputs from Montagu}
\usage{
update_coverage_values(site, iso3c, coverage_data, scenario_name)
}
\arguments{
\item{site}{site data file}

\item{iso3c}{country to update coverage values for}

\item{coverage_data}{VIMC vaccine forecast for site of interest}

\item{scenario_name}{scenario for vaccine forecast}
}
\value{
site file with variables 'rtss_cov', 'rtss_booster', 'r21_cov', 'r21_booster'
}
\description{
update vaccine coverage based on VIMC inputs from Montagu
}
