% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{vimc_postprocess}
\alias{vimc_postprocess}
\title{VIMC postprocessing}
\usage{
vimc_postprocess(
  output,
  le,
  iso3c,
  site_data,
  site_name,
  ur,
  vimc_pop,
  quick_run
)
}
\arguments{
\item{output}{model output}

\item{le}{VIMC life expectancy input}

\item{iso3c}{countrycode}

\item{site_data}{site file}

\item{site_name}{name of site}

\item{ur}{urbanicity}

\item{vimc_pop}{VIMC population input for country of interest}

\item{quick_run}{quick_run}
}
\description{
VIMC postprocessing
}
