% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimcmalaria-package.R
\docType{package}
\name{vimcmalaria-package}
\alias{vimcmalaria}
\alias{vimcmalaria-package}
\title{vimcmalaria: VIMC malaria modelling}
\description{
Helper functions for VIMC malaria modelling.
}
\author{
\strong{Maintainer}: Lydia Haile \email{l.haile@imperial.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
