% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{completed_reports}
\alias{completed_reports}
\title{Return a list of the reports that have already completed from orderly metadata}
\usage{
completed_reports(report_name)
}
\arguments{
\item{report_name}{name of orderly report to extract metadata from}
}
\description{
Return a list of the reports that have already completed from orderly metadata
}
