% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{completed_reports}
\alias{completed_reports}
\title{Pull completed reports based on orderly documentaiton}
\usage{
completed_reports(report_name)
}
\arguments{
\item{report_name}{name of orderly report to extract metadata from}
}
\description{
Pull completed reports based on orderly documentaiton
}
