% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{format_outputs}
\alias{format_outputs}
\title{format outputs for submission}
\usage{
format_outputs(dt, site_name, ur)
}
\arguments{
\item{dt}{postprocessed output}

\item{site_name}{name of site}

\item{ur}{urbanicity}
}
\description{
format outputs for submission
}
