% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{format_outputs}
\alias{format_outputs}
\title{format outputs for submission}
\usage{
format_outputs(
  dt,
  iso3c,
  site_name,
  ur,
  scenario,
  gfa,
  description,
  parameter_draw
)
}
\arguments{
\item{dt}{postprocessed output}

\item{iso3c}{country code}

\item{site_name}{name of site}

\item{ur}{urbanicity}

\item{scenario}{vaccine scenario}

\item{gfa}{global fund assumptions for other interventions (boolean)}

\item{description}{reason for model run}

\item{parameter_draw}{parameter draw}
}
\description{
format outputs for submission
}
