% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{process_output}
\alias{process_output}
\title{postprocess model output at the site level}
\usage{
process_output(
  model,
  le,
  site_data,
  vimc_pop,
  pop_single_yr,
  site_name,
  ur,
  scenario
)
}
\arguments{
\item{model}{model output}

\item{site_name}{site name}

\item{scenario}{vaccine scenario}
}
\value{
list of parameters for all model runs
}
\description{
postprocess model output at the site level
}
