% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{process_output}
\alias{process_output}
\title{postprocess model output at the site level}
\usage{
process_output(
  model,
  vimc_input,
  site_data,
  site_name,
  ur,
  iso3c,
  scenario,
  parameter_draw,
  quick_run,
  description
)
}
\arguments{
\item{model}{model output}

\item{vimc_input}{input file from VIMC}

\item{site_data}{site file for country of interest}

\item{site_name}{site name}

\item{ur}{urbanicity}

\item{iso3c}{country code}

\item{scenario}{vaccine scenario}

\item{parameter_draw}{parameter draw}

\item{quick_run}{quick run setting (boolean)}

\item{description}{description of reason for model run}
}
\value{
list of parameters for all model runs
}
\description{
postprocess model output at the site level
}
