% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R, R/postprocess.R
\name{pull_doses_output}
\alias{pull_doses_output}
\title{Aggregate VIMC outputs up to the country level}
\usage{
pull_doses_output(raw_output, processed_output)

pull_doses_output(raw_output, processed_output)
}
\arguments{
\item{raw_output}{model output}

\item{processed_output}{name of site}
}
\value{
data frame with doses of vaccine delivered per year
}
\description{
Aggregate VIMC outputs up to the country level

pull dosage output
}
