% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{pull_input_params}
\alias{pull_input_params}
\title{Paraemterize site + urbanicty of interest}
\usage{
pull_input_params(
  site_name,
  ur,
  site_data,
  coverage_data,
  scenario,
  parameter_draw,
  quick_run
)
}
\arguments{
\item{site_name}{name of site}

\item{ur}{urbanicity}

\item{site_data}{site file}

\item{coverage_data}{VIMC vaccine forecast for site of interest}

\item{scenario}{vaccine forecast scenario}
}
\value{
site file with additional variables 'rtss_coverage', 'rtss_booster_coverage', 'r21_coverage', 'r21_booster_coverage'
}
\description{
Paraemterize site + urbanicty of interest
}
