% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{recalibrate}
\alias{recalibrate}
\title{Very basic recalibration function (for Ethiopia sites)}
\usage{
recalibrate(params, site_name, site_dt)
}
\arguments{
\item{params}{simulation parameters}

\item{site_name}{name of site to recalibrate}

\item{site_dt}{site data}
}
\value{
recalibrated site
}
\description{
Very basic recalibration function (for Ethiopia sites)
}
