% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{scale_population}
\alias{scale_population}
\title{scale site file population to be consistent with country population input from VIMC}
\usage{
scale_population(site_data, site_name, ur, iso3c, vimc_pop, pop_single_yr)
}
\arguments{
\item{site_data}{site file}

\item{site_name}{name of site}

\item{ur}{urbanicity}

\item{iso3c}{country of interest}

\item{vimc_pop}{VIMC population input}

\item{pop_single_yr}{single-year age group VIMC population input}
}
\description{
scale site file population to be consistent with country population input from VIMC
}
