% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{update_coverage_values}
\alias{update_coverage_values}
\title{update vaccine coverage based on VIMC inputs from Montagu}
\usage{
update_coverage_values(site, coverage_data, scenario_name)
}
\arguments{
\item{site}{site data file}

\item{coverage_data}{VIMC vaccine forecast for site of interest}

\item{scenario_name}{scenario for vaccine forecast}
}
\value{
site file with additional variables 'rtss_coverage', 'rtss_booster_coverage', 'r21_coverage', 'r21_booster_coverage'
}
\description{
update vaccine coverage based on VIMC inputs from Montagu
}
