% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{vimc_postprocess}
\alias{vimc_postprocess}
\title{VIMC postprocessing}
\usage{
vimc_postprocess(
  output,
  le,
  iso3c,
  site_data,
  site_name,
  ur,
  vimc_pop,
  pop_single_yr
)
}
\arguments{
\item{output}{model output}

\item{le}{VIMC life expectancy input}

\item{iso3c}{countrycode}

\item{site_data}{site file}

\item{site_name}{name of site}

\item{ur}{urbanicity}

\item{pop_single_yr}{VIMC population input (single year age groups)}

\item{vimc_population}{VIMC population input}
}
\description{
VIMC postprocessing
}
